+++ tooling/mcerepkg/internal/rukpak/README.md
@@ -1,0 +1,3 @@
+# Source
+
+proudly stolen from <https://github.com/operator-framework/operator-controller/tree/main/internal/rukpak>
